macosx: Fix album duration type

+++ modules/gui/macosx/library/VLCLibraryDataTypes.h
@@ -200,7 +200,7 @@ typedef NS_ENUM(NSUInteger, VLCMediaLibraryParentGroupType) {
 @property (readonly) NSString *artistName;
 @property (readonly) NSString *genreString; // Lazy loaded for performance
 @property (readonly) int64_t artistID;
-@property (readonly) unsigned int duration;
+@property (readonly) int64_t duration;
 @property (readonly) unsigned int year;
 
 @end
